Rename order_formatting to datastructures and place under exchanges
