+++ rotkehlchen/exchanges/__init__.py
@@ -1,6 +1,6 @@
 from rotkehlchen.exchanges.binance import Binance  # noqa: F401
 from rotkehlchen.exchanges.bitmex import Bitmex  # noqa: F401
 from rotkehlchen.exchanges.bittrex import Bittrex  # noqa: F401
-from rotkehlchen.exchanges.exchange import Exchange, ExchangeInterface  # noqa: F401
+from rotkehlchen.exchanges.exchange import ExchangeInterface  # noqa: F401
 from rotkehlchen.exchanges.kraken import Kraken  # noqa: F401
 from rotkehlchen.exchanges.poloniex import Poloniex  # noqa: F401
