+++ rotkehlchen/tests/test_utils.py
@@ -3,8 +3,8 @@ import json
 import pytest
 
 from rotkehlchen.errors import UnprocessableTradePair
+from rotkehlchen.exchanges.data_structures import invert_pair
 from rotkehlchen.fval import FVal
-from rotkehlchen.order_formatting import invert_pair
 from rotkehlchen.serializer import process_result
 from rotkehlchen.utils.misc import combine_dicts, combine_stat_dicts, iso8601ts_to_timestamp
