+++ tools/data_faker/data_faker/actions.py
@@ -4,9 +4,9 @@ from typing import Tuple
 
 from rotkehlchen.assets.asset import Asset
 from rotkehlchen.constants.assets import A_BTC, A_EUR, A_USD, FIAT_CURRENCIES
+from rotkehlchen.exchanges.data_structures import Trade, TradeType, pair_get_assets
 from rotkehlchen.fval import FVal
 from rotkehlchen.history import PriceHistorian
-from rotkehlchen.order_formatting import Trade, TradeType, pair_get_assets
 from rotkehlchen.typing import Timestamp, TradePair
 
 STARTING_TIMESTAMP = 1464739200  # 01/06/2016
